% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codesymnmf_fit.R
\name{codesymnmf_fit}
\alias{codesymnmf_fit}
\title{Solves min_{H >0} \|A - HH'\|_F^2}
\usage{
codesymnmf_fit(A, H, maxiter = 100, rand_permutation = FALSE)
}
\arguments{
\item{A}{An n-by-n symmetric matrix}

\item{H}{An initial value for the n-by-r matrix H}

\item{maxiter}{The maximum number of iterations}

\item{rand_permutation}{A True/False value indicating whether to randomly
permute the columns of H when running coordinate descent}
}
\value{
A list containing the estimate for H and the objective function value
}
\description{
Solves min_{H >0} \|A - HH'\|_F^2
}
