% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codesymnmf_init.R
\name{codesymnmf_init}
\alias{codesymnmf_init}
\title{The initialization function for CoDesymNMF}
\usage{
codesymnmf_init(A, r, init_type)
}
\arguments{
\item{A}{An n-by-n symmetric matrix}

\item{r}{The number of columns in H}

\item{init_type}{The type of initialization to use for H.
If \code{init_type = 'zeros'}, then H is set to an n-by-r matrix of zeros.
Otherwise, H is set to an n-by-r matrix of entries drawn from standard uniform.}
}
\value{
Return an initial value for H
}
\description{
The initialization function for CoDesymNMF
}
